Fixed: System.ArgumentOutOfRangeException: Value of '20480' is not valid for 'Value'. 'Value' should be between 'Minimum' and 'Maximum'. trackWatermarkBackgroundTrans value load fixed.

+++ ZScreen/Global/FileSystem.cs
@@ -89,7 +89,7 @@ namespace ZSS
             {
                 img = WatermarkMaker.GetImage(img);
 
-                long size = Program.conf.SwitchAfter * 1024;
+                long size = (long)Program.conf.SwitchAfter * 1024;
 
                 MemoryStream ms = new MemoryStream();
                 MyGraphics.SaveImageToMemoryStream(img, ms, mImageFormats[Program.conf.FileFormat]);
